Avoid NullPointerException.

+++ src/net/sourceforge/fenixedu/domain/Site.java
@@ -73,7 +73,9 @@ public abstract class Site extends Site_Base {
 	    section.delete();
 	}
 
-	getAvailabilityPolicy().delete();
+	if (hasAvailabilityPolicy()) {
+	    getAvailabilityPolicy().delete();
+	}
 	removeCreator();
 	removeInitialContent();
 	removePortal();
